v17.0.4
- Critical bugfix for performance on MPEGTS streams
- Minor bugfixes and improvements

+++ www/nodejs-project/modules/streamer/adapters/ts.js
@@ -1,5 +1,5 @@
 const path = require('path'), StreamerAdapterBase = require('./base.js')
-const Downloader = require('../utils/downloader.js')
+const Downloader = require('../utils/downloader.js'), Joiner = require('../utils/joiner.js')
 
 class StreamerAdapterTS extends StreamerAdapterBase {
 	constructor(url, opts, cb){
@@ -18,18 +18,22 @@ class StreamerAdapterTS extends StreamerAdapterBase {
 					reject()
 				}
 			})
-			let args = []
+			const args = [this.url, this.opts]
 			if(global.config.get('ts-packet-filter-policy') == -1){
-				this.source = new Downloader(this.url, this.opts)
+				this.source = new Downloader(...args)
 			} else {
-				const MultiWorker = require('../../multi-worker')
-				this.worker = new MultiWorker()
-				this.source = this.worker.load(path.join(__dirname, '../utils/joiner-worker'))
-				args = [this.url, this.opts]
-				this.once('destroy', () => {
-					this.source && this.source.terminate()
-					this.worker && this.worker.terminate()
-				})
+				const useWorker = false
+				if(useWorker) {
+					const MultiWorker = require('../../multi-worker')
+					this.worker = new MultiWorker()
+					this.source = this.worker.load(path.join(__dirname, '../utils/joiner-worker'))
+					this.once('destroy', () => {
+						this.source && this.source.terminate()
+						this.worker && this.worker.terminate()
+					})
+				} else {
+					this.source = new Joiner(...args)
+				}
 			}
 			this.server = false
 			this.connectable = false
